refactor: Add CardinalDirection enum
This enum is useful / necessary to make routines that manipulate
directions type safe.

The fundamental issue is that the DirectionType enum includes the
variant "CURRENT", which cannot be decoded into a cardinal direction
(i.e. WEST) without knowledge of the Agent's prior cardinal direction.

This patch reintroduces a cartinal direction enum and includes basic
routines for manipulating cardinal directions, such as `rightOf`,
`leftOf`, and `reverseOf`.