feat: Add basic wall handling routine
This routine is "dumb" and intended to be called last, after the agent
has made an "intelligent" decision about the direction it should move.

If the agent has selected a blocked direction, this routine applies a
default sequence of turn right, turn left, go backwards in order to
navigate past blocked areas. See the docstring for more details.