fix: Make Controller member variables non-static
static member variables in Java are shared across all instances of the
class, which isn't what we want in this case since every Controller
instance should have its own data.

Also:

chore: Remove unnecessary temp variable

`routes.get(i)` returns a "reference" to the arraylist in the map, so
we can modify it directly without needing to make a copy.